Reallocated ct tests to ct-completion.test.ts.
