+++ tests/unit/python-utilities-valid.test.ts
@@ -1,38 +1,14 @@
 import process from "node:process";
-import * as nodeFetch from "node-fetch";
-import URI from "vscode-uri";
 import { describe, expect, test, vi, beforeEach, afterEach } from "vitest";
 import {
-  getChoiceData,
   updateVPythonCommand,
   getVPythonCommand,
-  choiceKindType,
   generateTestScriptDiagnostic,
-  generateCodedTestDiagnostic,
 } from "../../server/pythonUtilities";
 import path from "node:path";
-import { setServerState } from "../../src-common/vcastServer";
-import { asCompletionParameters, getCompletionPositionForLine } from "./utils";
-import {
-  Diagnostic,
-  DiagnosticSeverity,
-  TextDocument,
-} from "vscode-languageserver";
-import { getEnviroNameFromTestScript } from "../../server/serverUtilities";
-import { getCodedTestCompletionData } from "../../server/ctCompletions";
+import { Diagnostic, DiagnosticSeverity } from "vscode-languageserver";
 
 const timeout = 30_000; // 30 seconds
-
-vi.mock("node-fetch", async () => {
-  const actual = await vi.importActual<typeof nodeFetch>("node-fetch");
-
-  return {
-    ...actual,
-    default: vi.fn(),
-  };
-});
-
-const fetch = vi.mocked(nodeFetch.default);
 
 // Mock the child_process module
 vi.mock("child_process", async () => {
@@ -49,36 +25,6 @@ vi.mock("child_process", async () => {
       ),
   };
 });
-
-const expectedReceivedData = [
-  {
-    additionalTextEdits: [
-      {
-        newText: "// some extra data",
-        range: {
-          end: {
-            character: 0,
-            line: 0,
-          },
-          start: {
-            character: 0,
-            line: 0,
-          },
-        },
-      },
-    ],
-    data: 0,
-    detail: "",
-    kind: 1,
-    label: "unit",
-  },
-  {
-    data: 1,
-    detail: "",
-    kind: 1,
-    label: "Prototype-Stubs",
-  },
-];
 
 describe("Testing pythonUtilities (valid)", () => {
   beforeEach(() => {
@@ -99,122 +45,7 @@ describe("Testing pythonUtilities (valid)", () => {
     timeout
   );
 
-  test(
-    "validate getChoiceDataFromServer",
-    async () => {
-      const pythonUtilities = await import("../../src-common/vcastServer");
-
-      // Spy on `transmitCommand` and mock its implementation
-      vi.spyOn(pythonUtilities, "transmitCommand").mockResolvedValue({
-        success: true,
-        returnData: {
-          data: {
-            choiceKind: "File",
-            choiceList: ["unit", "Prototype-Stubs"],
-            extraText: "some extra data",
-            messages: ["some", "messages"],
-          },
-        },
-        statusText: "success",
-      });
-
-      setServerState(true);
-
-      // Need dummy coded test file for function (can be empty)
-      const unitTst = ``;
-
-      const lineToComplete = "// vmock";
-      const completionPosition = getCompletionPositionForLine(
-        lineToComplete,
-        unitTst
-      );
-
-      const envName = "vcast";
-
-      const languageId = "cpp";
-
-      const testEnvPath = path.join(
-        process.env.PACKAGE_PATH as string,
-        "tests",
-        "unit",
-        envName
-      );
-      const tstFilepath = path.join(
-        testEnvPath,
-        process.env.TST_FILENAME as string
-      );
-
-      const triggerCharacter = lineToComplete.at(-1);
-      const uri = URI.file(tstFilepath).toString();
-      const textDocument = TextDocument.create(uri, languageId, 1, unitTst);
-
-      const completion = asCompletionParameters(
-        textDocument,
-        completionPosition,
-        triggerCharacter
-      );
-
-      const enviroPath = getEnviroNameFromTestScript(tstFilepath);
-
-      if (enviroPath) {
-        const result = await getCodedTestCompletionData(
-          undefined,
-          lineToComplete,
-          completion,
-          enviroPath
-        );
-        expect(result).toEqual(expectedReceivedData);
-      }
-    },
-    timeout
-  );
-
-  test(
-    "validate getChoiceDataFromServer if it fails",
-    async () => {
-      // Mock fetch to simulate a failure and throw an error
-      fetch.mockImplementationOnce(() =>
-        Promise.reject(new Error("Failed to fetch: reason: Server down"))
-      );
-
-      // Set server state to true
-      setServerState(true);
-
-      const lineToComplete = "// vmock";
-      const envName = "vcast";
-
-      const testEnvPath = path.join(
-        process.env.PACKAGE_PATH as string,
-        "tests",
-        "unit",
-        envName
-      );
-      const tstFilepath = path.join(
-        testEnvPath,
-        process.env.TST_FILENAME as string
-      );
-
-      const enviroPath = getEnviroNameFromTestScript(tstFilepath);
-      let result: any;
-      if (enviroPath) {
-        result = await getChoiceData(
-          choiceKindType.choiceListCT,
-          enviroPath,
-          lineToComplete
-        );
-      }
-
-      expect(result).toStrictEqual({
-        choiceKind: "",
-        choiceList: [],
-        extraText: "",
-        messages: [],
-      });
-    },
-    timeout
-  );
-
-  test("should create and send a diagnostic objects for tst and ct", () => {
+  test("should create and send a diagnostic objects for tst", () => {
     const diagnostic: Diagnostic = {
       severity: DiagnosticSeverity.Warning,
       range: {
@@ -236,13 +67,6 @@ describe("Testing pythonUtilities (valid)", () => {
       sendDiagnostics: mockSendDiagnostics,
     };
 
-    // Functions under test
-    generateCodedTestDiagnostic(
-      connection,
-      "Test message",
-      "file:///path/to/document",
-      1
-    );
     generateTestScriptDiagnostic(
       connection,
       "Test message",
